Made staging of multiple files more robust
Previously via egg-status-buffer-stage-untracked-file you'd be able to
stage multiple files from a selected area, which would contain their
names. But if for example the following files where listed:

  aabc.txt
  xyz.txt

and you selected text from 'b' in the first line to 'z' in the second
line, the file names passed to git-add(1) would be: 'bc.txt' & 'xy'.

Now, this is no longer the case, the whole lines would be selected
and the area processed, so that 'aabc.txt' & 'xyz.txt' would be passed
to git-add(1).