Added a menu for Egg-Stash mode
The menu did not show up before at all when in Egg-Stash mode.
Enabled it and set the order of menu keybindings to appear as:

SPC:load details
h:hide/show details
a:apply
o:pop and apply stash
x:delete stash  X:delete all

The logic behind such arrangement is: first load to see the details
of the stash. Then apply it or pop & apply it. If not happy with the
stash delete it or delete all stashes.