fix cc-del and cc-add and improve navigation in status buffer
C-u n or p now will move to the next/previous block at the same level.
E.g. if the cursor is on a diff header, then C-u n will move to the next
diff header instead of the next hunk.

correctly highlighting cc-del and cc-add in log commit's details