fix: support slashes in branch/tag names
In GitFlow model we constantly use branch names containing slashes.
E.g. "refs/remotes/origin/feature/important-improvement". Egg used
incorrectly parse the such refs. After that fix "slashy" remote refs
displayed correctly in egg-log buffer. But surely there are many places
in the Egg code where it is understood that the name of the branch
is completely located after the last slash in: this too should
be investigated in future.